+++ Valorant Project/ValorantGUI/Windows/GameLogs/EditGameLog.xaml.cs
@@ -60,7 +60,7 @@ namespace ValorantGUI
                 int.Parse(KillsTextBox.Text.Trim()),
                 int.Parse(DeathsTextBox.Text.Trim()),
                 int.Parse(AssistsTextBox.Text.Trim()),
-                int.Parse(ADRTextBox.Text.Trim()),
+                float.Parse(ADRTextBox.Text.Trim()),
                 ((DateTime)DatePlayedPicker.SelectedDate).AddHours(_oldTime.Hour).AddMinutes(_oldTime.Minute).AddSeconds(_oldTime.Second));
             _gameLogPage.GameManager.UpdateGame(_selectedGame, args);
             _gameLogPage.PopulateGames();
